update chef recipes for demo
git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi/V_5.x/Server/trunk@25383 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ ChefCookbooks/installation.sh
@@ -1,20 +1,26 @@
 #!/bin/bash
-#SpagoBI Chef installation script
 set -e
 set -x
 
 #install chef 12 client
-wget https://opscode-omnibus-packages.s3.amazonaws.com/ubuntu/10.04/x86_64/chef_12.5.1-1_amd64.deb
-sudo dpkg -i chef_12.5.1-1_amd64.deb
-rm -f chef_12.5.1-1_amd64.deb
+if [ ! -d "/opt/chef" ]; then
+	wget https://opscode-omnibus-packages.s3.amazonaws.com/ubuntu/10.04/x86_64/chef_12.5.1-1_amd64.deb
+	sudo dpkg -i chef_12.5.1-1_amd64.deb
+	rm -f chef_12.5.1-1_amd64.deb
+fi
 
 #download recipes
-sudo apt-get update -q
-sudo apt-get -y install unzip
-wget https://github.com/SpagoBILabs/SpagoBI/releases/download/fiware-v5.1-1feb2d97af/chef-cookbooks.zip
-unzip chef-cookbooks.zip
-rm -f chef-cookbooks.zip
+sudo apt-get -y install subversion
+rm -rf cookbooks
+svn checkout https://github.com/SpagoBILabs/SpagoBI/trunk/ChefCookbooks/ cookbooks
+
 
 #install spagobi
-sudo chef-client -z -o 'recipe[spagobi::1.0.2_install]'
+if [ "$1" == "demo" ]; then
+	echo '{"demo":true}' | sudo chef-client -z -o 'recipe[spagobi::1.0.2_install]' -j /dev/stdin
+else
+	sudo chef-client -z -o 'recipe[spagobi::1.0.2_install]'
+fi
+
+#remove recipes
 rm -rf cookbooks
